Modified functions to convert contour/ellipse conversion to WCS

+++ src/sproc/Source.h
@@ -606,8 +606,8 @@ class Source : public Blob {
 		/**
 		* \brief Get fit ellipses
 		*/
-		int GetFitEllipses(std::vector<TEllipse*>& fitEllipses,bool useFWHM=true,bool convertToWCS=false,WCS* wcs=0,int coordSystem=-1,int pixOffset=0,bool useWCSSimpleConversion=true);
-	
+		int GetFitEllipses(std::vector<TEllipse*>& fitEllipses,bool useFWHM=true,bool convertToWCS=false,WCS* wcs=0,int coordSystem=-1,double pixOffset=0,bool useWCSSimpleConversion=true);
+
 		
 		/**
 		* \brief Get number of fit components
